openafs.spec: Use more familiar server paths
Currently the spec file uses paths in /var/openafs/ for many server
files (logs, server state, etc). However, it seems more common with
other packages, and for openafs on other platforms, to use
/var/lib/openafs for various states, and /var/log/openafs for logs.
So, use those paths instead.

As such, logs now go in /var/log/openafs, and we now use the dirs
/var/lib/openafs/db, /var/lib/openafs/local, and
/var/lib/openafs/backup. BosConfig still goes in
/etc/openafs/BosConfig.

Specify these new dirs in our configure invocation, and change all
relevant paths in the spec file. Remove our KRB5_CONFIG env var while
we're here, since the macro for its value, %{krb5config} is no longer
defined anywhere. Thus, without any --define arguments to rpmbuild,
this expands to the literal string "%{krb5config}", which is not
really what we want.

Also remove various files marked as %ghost and/or %config(noreplace)
that we never interact with (for example, server ThisCell,
rxkad.keytab or krb.conf). The RPM packaging never has to deal with
these files, as they are completely configured by the user, so there
is no reason to burden the packaging with them. The only benefit I can
see to marking such files as %ghost and %config(noreplace) is that RPM
is aware that the files belong to openafs; but, since they are in an
openafs-specific directory, that should already be obvious.