openafs.spec: Consolidate systemd check into macro
Instead of specifying '%if 0%{?fedora} >= 15 || 0%{?rhel >= 7}' in a
dozen places, just put that conditional in one place, to define a new
macro 'use_systemd'. This makes the conditionals around systemd files
a bit clearer, and makes it much easier to update the conditional if
we ever need to.

Also, always put the new systemd case first, before the sysvinit case,
just for consistency.