openafs.spec: Remove unversioned Obsoletes:
Currently the openafs-compat package is marked as Obsoletes:
openafs-client-compat. rpmlint complains that this is an unversioned
Obsoletes:, since this means we will obsolete any package called
openafs-client-compat, even if a new openafs-client-compat package is
created in the future.

All known potential openafs-client-compat packages are very old. The
newest ones I can find are from OpenAFS 1.2.2, from 2001, for Red Hat
6.2, 7.1, and 7.2 (that's pre-RHEL Red Hat), for example:
<http://www.openafs.org/dl/openafs/openafs/1.2.2/redhat-7.2/>

It may be impossible (and even if possible, very unlikely) that any
system using these RPMs would also have an openafs-client-compat
package that's over a decade old. And even if they did, lacking this
Obsoletes: would not cause a big problem (just a confusing error). So
just remove the Obsoletes: from openafs-compat to get rid of the
rpmlint noise and reduce the cruft in the spec file.