openafs.spec: Specify mode for ghosted CellServDB
Our default attributes for files are:

%defattr(-,root,root)

Which means we set the mode to whatever the mode is on disk, by
default. Our generated CellServDB is marked as a %ghost file, so it
doesn't actually exist on disk for us to install. This apparently
means it gets added to the RPM with a mode of 000, which makes rpmlint
complain that the modes are weird (since nobody can even read the
file).

To avoid this, give an explicit mode to the ghosted CellServDB. 0644
should be fine.