openafs.spec: Reintroduce %{krb5config}
Commit bb750852 removed setting KRB5_CONFIG to %{krb5config}, since
%{krb5config} was expanding to the literal string "%{krb5config}".
However, setting KRB5_CONFIG during configure can be useful, so
reintroduce the macro, but conditionalize its use, and put some info
about it near the top of the spec in comments.

Specify this as a configure argument, instead of an environment
variable. The environment variable KRB5_CONFIG actually has an
overloaded meaning (it points to the krb5.conf file to use for
krb5-enabled programs), and so specifying this as a configure argument
avoids the small chance of causing other unexpected consequences.