openafs.spec: Relax some dependencies
The client and server pieces really do not _require_ that we have the
same version of the base 'openafs' package installed. The docs package
doesn't really require anything else, and neither do the library
packages.

A few packages, like the kpasswd package and the compat packages, are
very small packages that are just extensions of the client or server
packages. For these, it makes a bit more sense to be more restrictive
on the requirements, since they require portions of the other packages
to exist, and they're not really their own standalone piece of
software.

Also get rid of the perl requirements (we don't use them), and get rid
of the unused kmod_name definition.