openafs.spec: Avoid x bit on non-executable files
Our sysconfig configuration file and our systemd service files are not
executables or scripts, so don't install them with the x bit turned
on. rpmlint complains that our sysconfig file is an executable, but is
marked as a config file and lacks a shebang. Removing the x bit should
get rid of these warnings.