openafs.spec: Require paths in compat packages
Our various compat packages contain a bunch of symlinks that point to
targets that exist in other packages. Any time this happens, rpmlint
complains that we have a dangling symlink, since it is incapable of
detecting of the target actually exists in one of our prerequisite
packages.

We can avoid this if we also mark an explicit dependency on each
specific file. This makes the dangling-symlink warnings go away, and
if we make a mistake, it will become obvious when trying to install
the package.