openafs.spec: Allow client install without kmod
The kmod openafs spec file is undergoing development (as is this
spec). At least for now, allow installing openafs-client without any
kernel module package. The client will not work without an openafs
kernel module, obviously, but this allows the binaries and other
components to at least be installed without one.

In the future the openafs-client package may also help the user in
finding or building a package with the openafs kernel module, so being
able to install openafs-client without a kernel module could be
helpful.