openafs.spec: Shuffle files around
Move some files around, mostly commands and their manpages. Ensure
that manpages are in the same package as their relevant command.

Ensure that commands in openafs-client are only useful on a client,
commands in openafs-server are only useful on a server, and commands
that are useful on both or neither are in the plain openafs package.
There are some exceptions to this for client-related utilities, since
server-side transarc paths historically have had references to client
utilities (e.g. fs). Those commands stay in the base openafs package,
so our openafs-transarc-server package can avoid depending on
openafs-client.